Add unit testing
