+++ src/main/java/training/calculatorreview/Main.java
@@ -1,0 +1,12 @@
+package training.calculatorreview;
+
+import java.util.Scanner;
+
+public class Main {
+
+    public static void main(String[] args) {
+        System.out.println("Welcome to the calculator!");
+        System.out.println("==========================");
+        new Scanner(System.in);
+    }
+}
