+++ src/test/java/MathTest.java
@@ -1,0 +1,10 @@
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+public class MathTest {
+    @Test
+    public void isMathsReal() {
+        assertTrue(1+1==2, "either something has gone very wrong or your computer doesnt know basic maths");
+    }
+}
